+++ algo/src/sccs/mod.rs
@@ -142,7 +142,7 @@ impl<C: AsMut<[usize]> + AsRef<[usize]>> Sccs<C> {
             .as_mut()
             .par_iter_mut()
             .for_each(|node_component| *node_component = inv_perm[*node_component]);
-        sizes.sort_by(|&x, &y| y.cmp(&x));
+        sizes.par_sort_by(|&x, &y| y.cmp(&x));
         sizes
     }
 }